Refactor login screen to use MVP pattern
