+++ core/src/main/kotlin/no/ntnu/beardblaster/menu/LoginMenuScreen.kt
@@ -1,41 +1,27 @@
 package no.ntnu.beardblaster.menu
 
-import com.badlogic.gdx.Gdx
 import com.badlogic.gdx.graphics.OrthographicCamera
 import com.badlogic.gdx.graphics.g2d.SpriteBatch
 import com.badlogic.gdx.scenes.scene2d.ui.TextButton
-import kotlinx.coroutines.flow.collect
-import kotlinx.coroutines.launch
 import ktx.actors.onChange
 import ktx.actors.onClick
 import ktx.assets.async.AssetStorage
-import ktx.async.KtxAsync
-import ktx.log.debug
-import ktx.log.error
-import ktx.log.info
-import ktx.log.logger
 import ktx.scene2d.scene2d
 import ktx.scene2d.textButton
 import no.ntnu.beardblaster.BaseScreen
 import no.ntnu.beardblaster.BeardBlasterGame
 import no.ntnu.beardblaster.assets.Nls
-import no.ntnu.beardblaster.commons.State
-import no.ntnu.beardblaster.commons.user.User
 import no.ntnu.beardblaster.ui.*
 import no.ntnu.beardblaster.user.LoginDialog
 import no.ntnu.beardblaster.user.RegisterDialog
-import no.ntnu.beardblaster.user.UserAuth
-import no.ntnu.beardblaster.user.UserRepository
-import pl.mk5.gdx.fireapp.auth.GdxFirebaseUser
-
-private val LOG = logger<LoginMenuScreen>()
 
 class LoginMenuScreen(
     game: BeardBlasterGame,
     batch: SpriteBatch,
     assets: AssetStorage,
     camera: OrthographicCamera
-) : BaseScreen(game, batch, assets, camera) {
+) : BaseScreen(game, batch, assets, camera), LoginMenuPresenter.View {
+    private val presenter = LoginMenuPresenter(this, game)
     private lateinit var exitBtn: TextButton
     private lateinit var loginBtn: TextButton
     private lateinit var registerBtn: TextButton
@@ -64,24 +50,13 @@ class LoginMenuScreen(
     override fun setBtnEventListeners() {
         loginBtn.onClick { showLoginDialog() }
         registerBtn.onClick { showRegisterDialog() }
-        exitBtn.onClick { Gdx.app.exit() }
+        exitBtn.onClick { presenter.onExitBtnClick() }
     }
-
-    override fun update(delta: Float) {}
 
     private fun showLoginDialog() {
         LoginDialog().apply {
             okBtn.onChange {
-                if (!UserAuth().isLoggedIn() && isValid) {
-                    UserAuth().signIn(email, password)
-                        .then<GdxFirebaseUser> {
-                            Gdx.app.postRunnable { game.setScreen<MenuScreen>() }
-                        }
-                        .fail { message, _ ->
-                            LOG.error { message }
-                            errorLabel.setText(message)
-                        }
-                }
+                presenter.onLoginClick(isValid, email, password)
             }
         }.show(stage)
     }
@@ -89,31 +64,12 @@ class LoginMenuScreen(
     private fun showRegisterDialog() {
         RegisterDialog().apply {
             okBtn.onChange {
-                LOG.debug { "Creating new user: ($username, $email, $password)" }
-                UserAuth().createUser(email, password, username)
-                    .then<GdxFirebaseUser> {
-                        val user = User(username, id = it.userInfo.uid)
-                        KtxAsync.launch {
-                            UserRepository().create(user, "users").collect { state: State<User> ->
-                                when (state) {
-                                    is State.Success -> {
-                                        game.setScreen<MenuScreen>()
-                                    }
-                                    is State.Failed -> {
-                                        LOG.error { state.message }
-                                        errorLabel.setText(state.message)
-                                    }
-                                    is State.Loading -> {
-                                        LOG.info { "Creating user.." }
-                                    }
-                                }
-                            }
-                        }
-                    }.fail { message, _ ->
-                        LOG.error { message }
-                        errorLabel.setText(message)
-                    }
+                presenter.onRegisterClick(username, email, password)
             }
         }.show(stage)
+    }
+
+    override fun setErrorLabel(msg: String) {
+        errorLabel.setText(msg)
     }
 }
